Add get_reserved_tracing_core helper to irq_common
The helper function was referenced by irq_migration.rs but missing
from irq_common.rs after a revert. This adds the function that parses
SCHTEST_IRQ_RESERVE_TRACING_CORE environment variable and fixes the
type mismatch in the comparison.

+++ src/cases/irq_migration.rs
@@ -74,7 +74,7 @@ fn irq_migration_test() -> Result<()> {
         }
         // Skip the reserved tracing core if set
         if let Some(reserved_cpu) = reserved_tracing_cpu {
-            if first_ht.id() == reserved_cpu {
+            if first_ht.id() == reserved_cpu as i32 {
                 continue;
             }
         }
